+++ web/src/components/index.ts
@@ -1,2 +1,4 @@
 export * from './foundation';
 export * from './layout';
+export * from './molecules';
+export * from './organisms';